Bug Fixes / Changes
Added UI to show which player's turn it currently is, triangles now match the colours of their respective flies (red and blue), triangles shouldn't move when the player is not in board view and fixed bug where pieces would be moved when a player moved a piece over them. Attempted assign player controllers on start but was unsuccessful may go back to that if I have time.