Call out cheating
Players can now call out cheating using the X button on a Xbox controller or C on the keyboard. If a player is caught cheating then the board is restored to the point before the acorn event and the cheating player turn ends. If falsely accused the player calling out the cheat takes damage instead.